Store type hierarchy for type values
Suggested by Peter
Instead of storing the leaf value per subject, store the whole hierarchy. By doing so, the frequencies are summed up on the more general levels of the hierarchy.
For example, 40% Politicians and 50% Athletes in a CS will be representented as (90% Thing, 90% Agent, 90% Person, 50% Athlete, 40% Politician), resulting in label candidate "Person" when threshold is set to 80%.