Enable+improve S5 (merging by FK) and improve S2 (merging by common ancestor).
For rule S2 with common ancestor, a constraint about the specific level is added. It is to prevent the case happened in webcrawl data where  such popular class as <Thing> only consumes small amount of percentage.

For rule S5, a condition considering the ratio between number of references and the frequency of referred CS is added.