Add hierarchy information for ontology-based name regardless of where the name comes from.
The name can be assigned based on the type-value, ontology-class similarity or FK. However, if the name can be found in ontology-class, we build ontology hierarchy for that name.
This is to prevent the problem happened when most of the CS's from dbpsb dataset get the name from their type values. These type values are also ontology-class, but we do not build any hierarchy information for these CS's.

More strict on the merging rule using common ancestor, putting the importance score to 0.001