Add labels from attributes of ontology classes
Now, we only add the labels for dbpedia and good relation ontologies.
For other ontology classes, we simply add NULL value as the attribute label. In the future, we can get that information by revising the attribute labels of other ontologies. However, as we observed, using labels from ontology is not better than simply extracting the last part of the attribute URI.