Use smalled buffer size when streaming out the lib
The bootstrap is slower now, but uses a lot less memory, also safer. So being conservative here.
While there, added a few comments to explain what's going on.

Note: if the in stream is null, we can just load the unpacked file, but I wanted to stream it out as well, verifying the stream copy the lib and load behaviour every time. You might want to change that, if building + testing tends to take too long (due to the rather small buffer). Or just increase the buffer size again.