[LiDAR]: Refactor the reading of files
So far files were being read column-wise. This meant that if we wanted
to read 3 attributes (e.g. XYZ) we would need to open and close the file
3 times. In the case of big LAZ files this would become prohibitively
expensive. This commit changes the implementation to row-wise read,
meaning, that for each point we read all the attributes the user has
requested.