Bugfix: codegen, move to the new API to return multiple outputs
The codegen API mil/mal to return multiple variables when invoking a statement
was changed in the master branch. With the recent merge the new API was
propagated and introduced a mismatch with what done for the graph related
statements. This fix aligns the API for these statements with the one used
for the rest of the code.