QRW: DCE pass for both SYS.NEST and paths
Unify the same logic to mark & remove the unused attributes in the DCE
pass for both the SYS.NEST aggregate and the graph shortest path
operator. There might still be a latent and occasional bug where the
list of attributes for a scalar function is not NULL even when the
type is not a nested table. Currently I am not able to reproduce it
anymore though :/