Major cleanup of error handling and clients IO
- The client code works strictly through the stream library.
The console readline option is removed.
- Error handling has been stratified. The MAL block
contains a 'errors' field to act as a container for
exceptions that come from deep in the code, e.g. the typechecker.
Most routines that originally used showScripException(), i.e.
sent the message to the client channel directly, has been
turned into use the 'errors' field.
- Several APIs have been changed to return proper exception strings.
- The MALreader and MAL parser have been changed to become more strict.
The MAL reader grabs a single MAL statement from the stream and
passes it as a null terminated string to the parser.
The client record now contains a line and cursor for parsing
- Initialization errors are cast into direct messages to stderr
- The MAL import routines have been simplified.

Mtest does not run due to errors in sql_aggr_*.mal
and the dynamic loading errors identified now require:
-disable-fits --disable-netcdf --disable-lidar --disable-shp
--disable-rintegration --without-samtools