Define a new prompt and handle it in the client
This is the most basic handling: the client recognizes the prompt,
parses it, saves the file name in the handle and skips the rest of the
block.

The new prompt consists of:

the bytes \001\003
a single space
the length of the file path
a single space
the file path

This is subject to change as we understand the problem better. For
instance maybe we do not need the length of the file path.

Next steps are to figure out how to coordinate client and server when
actually transferring the file.